requirements.txt
replicate
python-dotenv
streamlit
requests
Pillow
urllib3
